GREAT SUCCESS TEAM

+++ NOS_Android_App/app/src/main/java/ceg4110/nos_android_app/PendingMenuFolder.java
@@ -17,20 +17,27 @@ import android.widget.Toast;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.FileReader;
 import java.io.IOException;
+import java.util.Hashtable;
 
 public class PendingMenuFolder extends AppCompatActivity {
 
     private String result1, mCurrentPhotoPath, mCurrentPhotoPath1 = "/storage/emulated/0/Android/data/ceg4110.nos_android_app/files/Pending";
-    private static final int readReqCode = 42;
+    private static final int readReqCode = 43;
     String TAG = "TheTag";
     ImageView image;
     private String[] result;
     Context mContext;
+    private Hashtable<String, String> results;
+    File dict;
+    FileOutputStream outputStream;
+    boolean wrong = false;
 
 
     @Override
@@ -42,6 +49,34 @@ public class PendingMenuFolder extends AppCompatActivity {
             mCurrentPhotoPath = getIntent().getStringExtra("photoPath");
         }
 
+        results = new Hashtable<>();
+        mContext = getApplicationContext();
+        dict = new File("/storage/emulated/0/Android/data/ceg4110.nos_android_app/files/History/dict");
+        if (dict.exists()) {
+            StringBuilder in = new StringBuilder();
+            try {
+                BufferedReader br = new BufferedReader(new FileReader(dict));
+                String line;
+                outputStream = new FileOutputStream(dict, true);
+
+                while ((line = br.readLine()) != null) {
+                    results.put(line, br.readLine());
+                }
+                br.close();
+            } catch (IOException e) {
+                Log.e(TAG, "Error: " + e.getLocalizedMessage());
+            }
+
+        }
+        else {
+            try {
+                dict.createNewFile();
+                outputStream = new FileOutputStream(dict, true);
+            } catch (Exception e) {
+                Log.e(TAG, "dict problem: " + e.getLocalizedMessage());
+            }
+        }
+
         pendingFile();
 
     }
@@ -51,12 +86,22 @@ public class PendingMenuFolder extends AppCompatActivity {
      * returns to the History folder.
      */
     public void onClickDelete(View view) {
-        File toDelete = new File(mCurrentPhotoPath);
-        toDelete.delete();
-        Intent intent = new Intent(this, PendingMenuFolder.class);
-        startActivity(intent);
+        if (mCurrentPhotoPath == null)
+            Toast.makeText(getApplicationContext(), "No photo selected!", Toast.LENGTH_LONG).show();
+        else {
+            File toDelete = new File(mCurrentPhotoPath);
+            if (!toDelete.exists())
+                Toast.makeText(getApplicationContext(), "Photo chosen from wrong directory!", Toast.LENGTH_LONG).show();
+            else {
+                toDelete.delete();
+                Intent intent = new Intent(this, PendingMenuFolder.class);
+                startActivity(intent);
+            }
+        }
     }
-
+    /*
+     * Initiates main menu class
+     */
     public void onClickMainMenu(View view) {
         Intent intent = new Intent(this, MainMenu.class);
         startActivity(intent);
@@ -105,7 +150,10 @@ public class PendingMenuFolder extends AppCompatActivity {
 
         }
 
-        //upload
+         /*
+          * This method is called when the assess button is pressed.
+          * This method creates a new Uploader object that sends the photo to the AI for assessment
+          */
         @SuppressLint("StaticFieldLeak")
         public void onClickAssess(View view) {
             final Uploader uploader = new Uploader();
@@ -123,11 +171,22 @@ public class PendingMenuFolder extends AppCompatActivity {
                         progressDialog.show();
                     }
 
+                    /*
+                     * Sends photo to AI by image path
+                     * gets back string array from uploader holding the results.
+                     */
                     @Override
                     protected Boolean doInBackground(Void... params) {
                         Log.i(TAG, "Entering Uploader");
-                        Log.i(TAG, "path: " + mCurrentPhotoPath);
+                        if (mCurrentPhotoPath == null)
+                            return false;
                         File temp = new File(mCurrentPhotoPath);
+                        if (!temp.exists()) {
+                            wrong = true;
+                            return false;
+                        }
+                        Log.i(TAG, "path: " + mCurrentPhotoPath);
+
                         Log.i(TAG, "Does it exist? " + temp.exists());
                         result1 = uploader.uploadFile("picture", mCurrentPhotoPath);
                         result = uploader.getAllResults();
@@ -137,7 +196,10 @@ public class PendingMenuFolder extends AppCompatActivity {
                         else
                             return true;
                     }
-
+                    /*
+                     * Upon a successful assessment, this method takes the user to the results window
+                     * and moves the photo to the history folder
+                     */
                     @Override
                     protected void onPostExecute(Boolean aBoolean) {
                         super.onPostExecute(aBoolean);
@@ -151,7 +213,12 @@ public class PendingMenuFolder extends AppCompatActivity {
                         }
                         else {
                             Log.i(TAG, "Upload failed");
-                            Toast.makeText(getApplicationContext(), "Upload error! Leaving photo in Pending folder", Toast.LENGTH_LONG).show();
+                            if(mCurrentPhotoPath == null)
+                                Toast.makeText(getApplicationContext(), "No photo to upload!", Toast.LENGTH_LONG).show();
+                            else if(wrong)
+                                Toast.makeText(getApplicationContext(), "Photo chosen from wrong directory!", Toast.LENGTH_LONG).show();
+                            else
+                                Toast.makeText(getApplicationContext(), "Upload error! Leaving photo in Pending folder", Toast.LENGTH_LONG).show();
 
                         }
                     }
@@ -160,7 +227,9 @@ public class PendingMenuFolder extends AppCompatActivity {
                 Toast.makeText(getApplicationContext(), "No Internet Connection", Toast.LENGTH_LONG).show();
             }
         }
-
+        /*
+         * This method initiates the results class and sends the photo path and results to the class
+         */
         public void goToResults(String path){
             if (!path.equals("")) {
                 Intent intent = new Intent(this, ResultScreen.class);
@@ -172,7 +241,9 @@ public class PendingMenuFolder extends AppCompatActivity {
                 Log.e(TAG, "results path problem, line 135");
             }
         }
-
+        /*
+         *This method moves a photo to the history folder
+         */
         private String movePhoto() {
             File photoFile = new File(mCurrentPhotoPath);
             String path = "";
@@ -233,11 +304,11 @@ public class PendingMenuFolder extends AppCompatActivity {
                 Log.e(TAG, "hate 3, " + e.getLocalizedMessage());
             }
 
-            /*
+
             results.put(mCurrentPhotoPath.substring(mCurrentPhotoPath.lastIndexOf('/')), result[1] + " " + result[2]);
 
             try {
-                outputStream.write((mCurrentPhotoPath.substring(mCurrentPhotoPath.lastIndexOf('/') + 1) + "\n").getBytes());
+                outputStream.write((mCurrentPhotoPath.substring(mCurrentPhotoPath.lastIndexOf('/')) + "\n").getBytes());
                 Log.i(TAG, "past path, writing results");
                 outputStream.write((result[1] + " " + result[2] + "\n").getBytes());
             } catch (IOException ex) {
@@ -245,7 +316,7 @@ public class PendingMenuFolder extends AppCompatActivity {
             } catch (Exception e) {
                 Log.e(TAG, "Error with dict: " + e.getLocalizedMessage());
             }
-            */
+
             photoFile.delete();
             return path;
         }
